pyproject.toml
[tool.poetry]
name = "ai-hedge-fund"
version = "0.1.0"
description = "An AI-powered hedge fund that uses multiple agents to make trading decisions"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "src", from = "." },
    { include = "app", from = "." }
]

[tool.poetry.dependencies]
python = "^3.11"
# Core Language Model Framework
langchain = "^0.3.7"
langchain-anthropic = "0.3.5"
langchain-groq = "0.2.3"
langchain-openai = "^0.3.5"
langchain-deepseek = "^0.1.2"
langchain-ollama = "0.3.6"
langgraph = "0.2.56"
langchain-google-genai = "^2.0.11"
langchain-gigachat = "^0.3.12"
# Scientific Computing Stack for Regime Detection and ML Ensemble
numpy = "^1.24.0"  # Core numerical computing
pandas = "^2.1.0"  # Data manipulation and analysis
scipy = "^1.11.0"  # Statistical functions for regime detection
hmmlearn = "^0.3.0"  # Hidden Markov Models for market regime classification

# Machine Learning Stack for Signal Ensemble (Phase 2)
scikit-learn = "^1.3.0"  # ML ensemble models and preprocessing
joblib = "^1.3.0"  # Model persistence and parallel processing
# Visualization and Output
matplotlib = "^3.9.2"  # Plotting for backtesting results
tabulate = "^0.9.0"  # Table formatting for output
colorama = "^0.4.6"  # Colored console output
rich = "^13.9.4"  # Rich text and progress bars
questionary = "^2.1.0"  # Interactive CLI prompts
# Configuration and Environment
python-dotenv = "1.0.0"  # Environment variable management
# Backend API Dependencies
fastapi = {extras = ["standard"], version = "^0.104.0"}
fastapi-cli = "^0.0.7"
pydantic = "^2.4.2"
httpx = "^0.27.0"
sqlalchemy = "^2.0.22"
alembic = "^1.12.0"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 420
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
force_alphabetical_sort_within_sections = true